+++ react/TextAreaBlock.tsx
@@ -29,6 +29,7 @@ import {
   fetchEmailTemplateLogo,
   bindTableData,
 } from './utils/excelUtils'
+import { isValidToAddItems } from './utils/checkBrandRestrictions'
 
 const messages = defineMessages({
   success: {
@@ -58,7 +59,8 @@ interface ItemType {
   quantity: number
 }
 
-const SPECAIL_BRAND_NAME = 'SPIRALOCK'
+// const SPECAIL_BRAND_NAME = 'SPIRALOCK'
+// const SPECAIL_BRAND_NAME_2 = 'SWS SPARES'
 
 const TextAreaBlock: FunctionComponent<
   TextAreaBlockInterface & WrappedComponentProps
@@ -308,38 +310,14 @@ const TextAreaBlock: FunctionComponent<
   const handles = useCssHandles(CSS_HANDLES)
 
   const addToCartCopyNPaste = () => {
-    const currentItemsInCart = orderForm.orderForm.items
-
-    const isSpecialBrandItemExistInCurrentCart = (
-      currentItemsInCart ?? []
-    ).find(
-      (item: any) =>
-        item?.additionalInfo?.brandName?.toUpperCase() === SPECAIL_BRAND_NAME
+    const validToAddItems = isValidToAddItems(
+      orderForm?.orderForm?.items?.map(
+        (item: any) => item?.additionalInfo?.brandName ?? ''
+      ),
+      reviewItems?.map((item: any) => item?.brand ?? '')
     )
 
-    const specialBrandItemInReviewItems = (reviewItems ?? []).filter(
-      (item: any) => item.brand.toUpperCase() === SPECAIL_BRAND_NAME
-    )
-
-    const cond1 =
-      currentItemsInCart.length > 0 &&
-      !!isSpecialBrandItemExistInCurrentCart &&
-      specialBrandItemInReviewItems.length === reviewItems.length
-
-    const cond2 =
-      currentItemsInCart.length === 0 &&
-      specialBrandItemInReviewItems.length === reviewItems.length
-
-    const cond3 =
-      currentItemsInCart.length === 0 &&
-      specialBrandItemInReviewItems.length === 0
-
-    const cond4 =
-      currentItemsInCart.length > 0 &&
-      !isSpecialBrandItemExistInCurrentCart &&
-      specialBrandItemInReviewItems.length === 0
-
-    if (cond1 || cond2 || cond3 || cond4) {
+    if (validToAddItems) {
       const items: any = reviewItems
         .filter((item: any) => item.error === null && item.vtexSku !== null)
         .map(({ vtexSku, quantity, seller, unit }: any) => {